refactor: create countries module
