+++ src/components/forms/club/clubs-filter-form.tsx
@@ -3,7 +3,7 @@ import { Field, Form, Formik } from 'formik'
 import { useTranslation } from 'next-i18next'
 
 import { ClubsFiltersDto } from '@/modules/clubs/types'
-import { CountryDto } from '@/types/countries'
+import { CountryDto } from '@/modules/countries/types'
 import { RegionDto } from '@/types/regions'
 
 import { CountriesCombo } from '../../selects/countries-combo'
