+++ src/components/forms/player/create-player.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'next-i18next'
 import { useAlertsState } from '@/context/alerts/useAlertsState'
 import { CreatePlayerDto } from '@/modules/players/types'
 import { TeamBasicDataDto } from '@/modules/teams/types'
-import { CountryDto } from '@/types/countries'
+import { CountryDto } from '@/modules/countries/types'
 import { PlayerPositionDto } from '@/types/player-positions'
 
 import { Container } from '../container'
