+++ src/pages/players/edit/[slug].tsx
@@ -5,9 +5,9 @@ import { ErrorContent } from '@/components/error/error-content'
 import { EditPlayerForm } from '@/components/forms/player/edit-player'
 import { Loader } from '@/components/loader/loader'
 import { PageHeading } from '@/components/page-heading/page-heading'
-import { useCountriesList } from '@/lib/countries'
 import { usePlayerPositionsList } from '@/lib/player-positions'
 import { withSessionSsr } from '@/lib/session'
+import { useCountriesList } from '@/modules/countries/hooks'
 import { useUpdatePlayer } from '@/modules/players/hooks'
 import { PlayerDto } from '@/modules/players/types'
 import { useTeamsList } from '@/modules/teams/hooks'
